sql homework 1.4

+++ src/developer/sql/1.sql
@@ -1,4 +1,4 @@
-ALTER TABLE developers ADD COLUMN salary INTEGER DEFAULT 0;
+ALTER TABLE developers ADD COLUMN salary NUMERIC(10,2) DEFAULT 0;
 
 UPDATE developers SET salary = 110000 WHERE id = 1;
 UPDATE developers SET salary = 120000 WHERE id = 2;
@@ -12,4 +12,4 @@ UPDATE developers SET salary = 190000 WHERE id = 9;
 UPDATE developers SET salary = 200000 WHERE id = 10;
 UPDATE developers SET salary = 210000 WHERE id = 11;
 UPDATE developers SET salary = 220000 WHERE id = 12;
-UPDATE developers SET salary = 1000 WHERE id = 13;
+UPDATE developers SET salary = 1000   WHERE id = 13;
